% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmnchfunding-package.R
\docType{package}
\name{rmnchfunding-package}
\alias{rmnchfunding}
\alias{rmnchfunding-package}
\title{rmnchfunding: Tools For Reproducible RMNCH Funding Analysis}
\description{
Calculates bilateral and multilateral funding for reproductive, maternal, newborn and child health (RMNCH), sexual and reproductive health and rights (SRHR) and family planning from OECD Creditor Reporting System (CRS) flows and OECD providers' total use of the multilateral system, using the revised Muskoka method. Ships the sector and multilateral-agency coefficient tables the method depends on, versioned by methodology vintage and adjustable by the user.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meltemod/rmnchfunding}
  \item \url{https://meltemod.github.io/rmnchfunding/}
  \item Report bugs at \url{https://github.com/meltemod/rmnchfunding/issues}
}

}
\author{
\strong{Maintainer}: Meltem Odabaş \email{me@meltemodabas.net} (\href{https://orcid.org/0000-0001-6123-0676}{ORCID})

Authors:
\itemize{
  \item Meltem Odabaş \email{me@meltemodabas.net} (\href{https://orcid.org/0000-0001-6123-0676}{ORCID})
}

}
\keyword{internal}
