% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmnchfunding-package.R
\docType{package}
\name{rmnchfunding-package}
\alias{rmnchfunding}
\alias{rmnchfunding-package}
\title{rmnchfunding: Tools For Reproducible RMNCH Funding Analysis}
\description{
Calculates bilateral and multilateral reproductive, maternal, newborn and child health (RMNCH) and family-planning funding from OECD Creditor Reporting System (CRS) data under three methodologies, with user-adjustable percentage tables. Provides muskoka1(), muskoka2() and revised_fp().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meltemod/rmnchfunding}
  \item \url{https://meltemod.github.io/rmnchfunding/}
  \item Report bugs at \url{https://github.com/meltemod/rmnchfunding/issues}
}

}
\author{
\strong{Maintainer}: Meltem Odabaş \email{me@meltemodabas.net} (\href{https://orcid.org/0000-0001-6123-0676}{ORCID})

Authors:
\itemize{
  \item Meltem Odabaş \email{me@meltemodabas.net} (\href{https://orcid.org/0000-0001-6123-0676}{ORCID})
}

}
\keyword{internal}
